+++ php/src/Application/Command/InvocationInfo.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace PhpIntegrator\Application\Command;
+
+use ArrayAccess;
+use UnexpectedValueException;
+
+use GetOptionKit\OptionCollection;
+
+/**
+ * Allows fetching invocation information of a method or function call.
+ */
+class InvocationInfo extends AbstractCommand
+{
+    /**
+     * @inheritDoc
+     */
+    protected function attachOptions(OptionCollection $optionCollection)
+    {
+        $optionCollection->add('file?', 'The file to examine.')->isa('string');
+        $optionCollection->add('stdin?', 'If set, file contents will not be read from disk but the contents from STDIN will be used instead.');
+        $optionCollection->add('charoffset?', 'If set, the input offset will be treated as a character offset instead of a byte offset.');
+        $optionCollection->add('offset:', 'The character byte offset into the code to use for the determination.')->isa('number');
+    }
+
+    /**
+     * @inheritDoc
+     */
+    protected function process(ArrayAccess $arguments)
+    {
+        if (!isset($arguments['file']) && (!isset($arguments['stdin']) || !$arguments['stdin']->value)) {
+            throw new UnexpectedValueException('Either a --file file must be supplied or --stdin must be passed!');
+        } elseif (!isset($arguments['offset'])) {
+            throw new UnexpectedValueException('An --offset must be supplied into the source code!');
+        }
+
+        $code = $this->getSourceCodeHelper()->getSourceCode(
+            isset($arguments['file']) ? $arguments['file']->value : null,
+            isset($arguments['stdin']) && $arguments['stdin']->value
+        );
+
+        $offset = $arguments['offset']->value;
+
+        if (isset($arguments['charoffset']) && $arguments['charoffset']->value == true) {
+            $offset = $this->getSourceCodeHelper()->getByteOffsetFromCharacterOffset($offset, $code);
+        }
+
+        $result = $this->getInvocationInfoAt($code, $offset);
+
+        return $this->outputJson(true, $result);
+    }
+
+    /**
+     * @param string $code
+     * @param int    $offset
+     *
+     * @return array
+     */
+    public function getInvocationInfoAt($code, $offset)
+    {
+        return $this->getInvocationInfo(substr($code, 0, $offset));
+    }
+
+    /**
+     * @param string $code
+     *
+     * @return array
+     */
+    public function getInvocationInfo($code)
+    {        
+        return $this->getSourceCodeHelper()->getInvocationInfoAt($code);
+    }
+}