+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * When retrieving available variables or deducing types, character offsets were not correctly being translated to byte offsets, sometimes resulting in incorrect results.
 
 ### Changes for developers
+* `getInvocationInfo` is now also available separately as PHP command.
+* `getInvocationInfoAt` will now return an `offset` rather than a `bufferPosition` to be consistent with the other commands.
 * `getResultingTypesAt` is now simply a convenience call to `deduceTypes` as its underlying code has been completely moved to PHP.
 * `deduceTypes` gained a new parameter `ignoreLastElement`, which does the same as the identically-named parameter for `getResultingTypesAt` does.